[EI-153] [web] refactor: Add hook folder for view components
