+++ web/app/ui/components/view/hooks/use-alert-dialog.hook.ts
