+++ web/app/ui/components/view/molocule/dialog-menu/index.ts
@@ -1,7 +1,7 @@
 import { DialogMenuRoot } from './dialog-menu-root';
 import { DialogMenuItem } from './dialog-menu-item';
 import { DialogMenuHeader } from './dialog-menu-header';
-export { useDialogMenu } from './use-dialog-menu.hook';
+export { useDialogMenu } from '../../hooks/use-dialog-menu.hook';
 
 const DialogMenu = Object.assign(DialogMenuRoot, {
   Header: DialogMenuHeader,
